requirements.txt
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
django-colorfield==0.13.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework-simplejwt
google-auth
requests
cryptography
python-decouple
gunicorn
